refactor: (#280) value 객체 Length & Range 유효성 검사
