+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/common/dto/request/CheckMatchedAccountWebRequest.kt
@@ -1,9 +1,9 @@
 package team.comit.simtong.domain.common.dto.request
 
-import team.comit.simtong.global.value.EmployeeNumber
+import org.hibernate.validator.constraints.Range
+import team.comit.simtong.domain.user.value.EmployeeNumber
 import javax.validation.constraints.Email
 import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotNull
 
 /**
  *
@@ -11,10 +11,10 @@ import javax.validation.constraints.NotNull
  *
  * @author Chokyunghyeon
  * @date 2022/10/15
- * @version 1.0.0
+ * @version 1.2.3
  **/
 data class CheckMatchedAccountWebRequest(
-    @field:NotNull
+    @field:Range(min = EmployeeNumber.MIN_VALUE, max = EmployeeNumber.MAX_VALUE)
     val employeeNumber: EmployeeNumber,
 
     @field:NotBlank
