+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/user/dto/request/ChangeNicknameWebRequest.kt
@@ -1,6 +1,7 @@
 package team.comit.simtong.domain.user.dto.request
 
-import team.comit.simtong.global.value.NickName
+import org.hibernate.validator.constraints.Length
+import team.comit.simtong.domain.user.value.NickName
 import javax.validation.constraints.Pattern
 
 /**
@@ -12,6 +13,7 @@ import javax.validation.constraints.Pattern
  * @version 1.2.3
  **/
 data class ChangeNicknameWebRequest(
-    @field:Pattern(regexp = NickName.PATTERN)
+    @Length(min = NickName.MIN_LENGTH, max = NickName.MAX_LENGTH)
+    @Pattern(regexp = NickName.PATTERN)
     val nickname: NickName
 )
