+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/user/dto/request/SignInWebRequest.kt
@@ -1,21 +1,22 @@
 package team.comit.simtong.domain.user.dto.request
 
-import team.comit.simtong.global.value.EmployeeNumber
-import team.comit.simtong.global.value.Password
+import org.hibernate.validator.constraints.Range
+import team.comit.simtong.domain.user.value.EmployeeNumber
+import team.comit.simtong.domain.user.value.Password
 import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotNull
 
 /**
  *
  * 일반 사용자가 로그인을 요청하는 SignInWebRequest
  *
  * @author kimbeomjin
+ * @author Chokyunghyeon
  * @date 2022/09/08
  * @version 1.2.3
  **/
 data class SignInWebRequest(
 
-    @field:NotNull
+    @field:Range(min = EmployeeNumber.MIN_VALUE, max = EmployeeNumber.MAX_VALUE)
     val employeeNumber: EmployeeNumber,
 
     @field:NotBlank
