+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/user/dto/request/SignUpWebRequest.kt
@@ -1,12 +1,12 @@
 package team.comit.simtong.domain.user.dto.request
 
+import org.hibernate.validator.constraints.Length
 import org.hibernate.validator.constraints.Range
-import team.comit.simtong.global.value.EmployeeNumber
-import team.comit.simtong.global.value.NickName
-import team.comit.simtong.global.value.Password
+import team.comit.simtong.domain.user.value.EmployeeNumber
+import team.comit.simtong.domain.user.value.NickName
+import team.comit.simtong.domain.user.value.Password
 import javax.validation.constraints.Email
 import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotNull
 import javax.validation.constraints.Pattern
 
 /**
@@ -26,17 +26,15 @@ data class SignUpWebRequest(
     @field:Email
     val email: String,
 
-    @field:NotNull
-    @field:Range(
-        min = EmployeeNumber.MIN,
-        max = EmployeeNumber.MAX
-    )
+    @Range(min = EmployeeNumber.MIN_VALUE, max = EmployeeNumber.MAX_VALUE)
     val employeeNumber: EmployeeNumber,
 
-    @field:Pattern(regexp = Password.PATTERN)
+    @Length(min = Password.MIN_LENGTH, max = Password.MAX_LENGTH)
+    @Pattern(regexp = Password.PATTERN)
     val password: Password,
 
-    @field:Pattern(regexp = NickName.PATTERN)
+    @Length(min = NickName.MIN_LENGTH, max = NickName.MAX_LENGTH)
+    @Pattern(regexp = NickName.PATTERN)
     val nickname: NickName,
 
     val profileImagePath: String?,
